simplefin should use formatPayeeName
