+++ src/app-gocardless/banks/american-express-aesudef1.js
@@ -1,8 +1,5 @@
-import {
-  amountToInteger,
-  sortByBookingDateOrValueDate,
-  formatPayeeName,
-} from '../utils.js';
+import { amountToInteger, sortByBookingDateOrValueDate } from '../utils.js';
+import { formatPayeeName } from '../../util/payee-name.js';
 
 /** @type {import('./bank.interface.js').IBank} */
 export default {
