+++ src/app-gocardless/banks/belfius_gkccbebb.js
@@ -1,5 +1,5 @@
 import Fallback from './integration-bank.js';
-import { formatPayeeName } from '../utils.js';
+import { formatPayeeName } from '../../util/payee-name.js';
 
 /** @type {import('./bank.interface.js').IBank} */
 export default {
