+++ src/app-gocardless/banks/integration-bank.js
@@ -3,8 +3,8 @@ import {
   sortByBookingDateOrValueDate,
   amountToInteger,
   printIban,
-  formatPayeeName,
 } from '../utils.js';
+import { formatPayeeName } from '../../util/payee-name.js';
 
 const SORTED_BALANCE_TYPE_LIST = [
   'closingBooked',
