+++ src/app-gocardless/banks/sparnord-spnodk22.js
@@ -2,8 +2,8 @@ import {
   sortByBookingDateOrValueDate,
   amountToInteger,
   printIban,
-  formatPayeeName,
 } from '../utils.js';
+import { formatPayeeName } from '../../util/payee-name.js';
 
 const SORTED_BALANCE_TYPE_LIST = [
   'closingBooked',
