+++ src/app-gocardless/banks/spk-karlsruhe-karsde66.js
@@ -2,8 +2,8 @@ import {
   printIban,
   amountToInteger,
   sortByBookingDateOrValueDate,
-  formatPayeeName,
 } from '../utils.js';
+import { formatPayeeName } from '../../util/payee-name.js';
 
 /** @type {import('./bank.interface.js').IBank} */
 export default {
